requirements_streamlit.txt
streamlit==1.28.1
langchain==0.0.335
chromadb==0.4.18
sentence-transformers==2.2.2
PyMuPDF==1.23.5
google-generativeai==0.3.2 